# Speedup exclude/include checking