What does this PR change? What problem does it solve?
-----------------------------------------------------
Backup (and restore) slow down a lot if a large number of excludes is used, see e.g. #2694. This PR optimizes the pattern processing to reduce the number and size of memory allocations. In total this speeds up the benchmarks by a factor between 5 and 100:

```
name                          old time/op    new time/op    delta
FilterPatterns/Relative-4       62.1ms ±15%    11.0ms ± 2%  -82.25%  (p=0.000 n=9+10)
FilterPatterns/Absolute-4        111ms ±10%       9ms ± 3%  -92.24%  (p=0.000 n=10+10)
FilterPatterns/Wildcard-4        393ms ±15%      36ms ± 2%  -90.73%  (p=0.000 n=10+9)
FilterPatterns/ManyNoMatch-4     10.0s ± 3%      0.1s ± 6%  -98.98%  (p=0.000 n=10+9)

name                          old alloc/op   new alloc/op   delta
FilterPatterns/Relative-4       16.4MB ± 0%     3.6MB ± 0%  -78.25%  (p=0.000 n=10+9)
FilterPatterns/Absolute-4       31.4MB ± 0%     3.6MB ± 0%  -88.62%  (p=0.000 n=9+10)
FilterPatterns/Wildcard-4        168MB ± 0%      20MB ± 0%  -88.25%  (p=0.000 n=10+10)
FilterPatterns/ManyNoMatch-4    3.23GB ± 0%    0.00GB ± 0%  -99.89%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
FilterPatterns/Relative-4         178k ± 0%       22k ± 0%  -87.50%  (p=0.000 n=10+10)
FilterPatterns/Absolute-4         266k ± 0%       22k ± 0%  -91.67%  (p=0.000 n=10+10)
FilterPatterns/Wildcard-4        1.87M ± 0%     0.09M ± 0%  -95.27%  (p=0.000 n=10+10)
FilterPatterns/ManyNoMatch-4     17.7M ± 0%      0.0M ± 0%  -99.88%  (p=0.000 n=9+10)
```

The`ManyNoMatch` test uses 200 exclude patterns which are suffixed such that these do not match any filenames, but require matching against the full pattern which is the worst-case performance wise. Originally I've tried to add all 22k tested filepaths to the filter patterns, however that was way too slow.

Assuming restic should be able to check 20k excludes per second (i.e., 1.2 million excludes per minute), one "op" of the `ManyNoMatch` benchmark shouldn't take much more than one seconds. With the optimizations I'm currently able to increase the number of exclude pattern to 2000.

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------
The exclude performance problems were in parts discussed in #2694.

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
- [x] I have added tests for all changes in this PR **The code already had a lot of tests, I've added some special case tests to increase coverage to 95%**
- ~~[ ] I have added documentation for the changes (in the manual)~~
- [ ] There's a new file in `changelog/unreleased/` that describes the changes for our users (template [here](https://github.com/restic/restic/blob/master/changelog/TEMPLATE))
- [x] I have run `gofmt` on the code in all commits
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [x] I'm done, this Pull Request is ready for review