@MichaelEischer 2020-10-07
I've fixed the "ManyNoMatch" (previously "Many") benchmark to actually work as intended and to cause the worst-case performance.