@MichaelEischer 2020-10-07
Profiling the optimized code now reports that more than half of the time is spent in `hasDoubleWildcard`. However, optimizing that function away just shifts the costs to some other part of the `match` function. So it looks like the string comparisons begin to be memory-bound.