@greatroar 2020-10-07
Impressive result. I implemented my glob-to-regexp translation idea, and while it gets rid of the allocations, it's almost twice as slow as the current code.