@greatroar 2020-10-09
MichaelEischer/restic#1 gives a further 25% speedup on the hardest benchmark.