@MichaelEischer 2020-10-09
@greatroar Wow, thanks for the extra speed-up. I've updated the comparison in the PR description, now we're at a speed-up by a smooth factor of 100. 🥳